+++ api/repository/types.go
@@ -11,8 +11,8 @@ type UserRepository interface {
 	FindUserByEmail(email string) (model.User, error)
 	InsertUserVerification(userId uint, code string) error
 	UpdateUserStatusByIdAndCode(userId uint, code string) (model.User, error)
-	FindUserById(userId uint) (model.User,error)
-	UpdateAmount(user model.User,amount float64) (model.User, error)
+	FindUserById(userId uint) (model.User, error)
+	UpdateAmount(user model.User, amount int64) (model.User, error)
 }
 
 type OrderRepository interface{
